--- src/main/java/kr/modernworld/modernworldv2/user/application/UserPointService.java
@@ -1,42 +1,0 @@
-package kr.modernworld.modernworldv2.user.application;
-
-import kr.modernworld.modernworldv2.global.error.BusinessErrorCode;
-import kr.modernworld.modernworldv2.global.error.BusinessException;
-import kr.modernworld.modernworldv2.user.domain.port.user.UserRepository;
-import kr.modernworld.modernworldv2.user.domain.user.User;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-public class UserPointService {
-
-  private final UserRepository userRepository;
-
-  @Transactional
-  public void decreaseCurrentPoint(Long userNo, Long price) {
-    User user = userRepository.findUserByUserNoForUpdate(userNo)
-        .orElseThrow(() -> new BusinessException(BusinessErrorCode.USER_NOT_FOUND));
-
-    try {
-      user.decreaseCurrentPoint(price);
-    } catch (IllegalArgumentException e) {
-      throw new BusinessException(BusinessErrorCode.USER_NOT_HAS_ENOUGH_POINT,
-          " " + e.getMessage());
-    }
-
-    userRepository.save(user);
-  }
-
-  @Transactional
-  public void increaseCurrentAccumulationPoint(Long userNo, Long price) {
-    User user = userRepository.findUserByUserNoForUpdate(userNo)
-        .orElseThrow(() -> new BusinessException(BusinessErrorCode.USER_NOT_FOUND));
-
-    user.increaseCurrentAccumulationPoint(price);
-
-    userRepository.save(user);
-  }
-
-}
