+++ src/main/java/kr/modernworld/modernworldv2/user/application/user/UserService.java
@@ -1,6 +1,10 @@
-package kr.modernworld.modernworldv2.user.application.oauth;
+package kr.modernworld.modernworldv2.user.application.user;
 
+import kr.modernworld.modernworldv2.global.error.BusinessErrorCode;
+import kr.modernworld.modernworldv2.global.error.BusinessException;
 import kr.modernworld.modernworldv2.user.application.LegendService;
+import kr.modernworld.modernworldv2.user.application.oauth.OAuthTokenDTO;
+import kr.modernworld.modernworldv2.user.application.oauth.SocialUserInfoDTO;
 import kr.modernworld.modernworldv2.user.domain.port.OAuthClient;
 import kr.modernworld.modernworldv2.user.domain.port.user.UserQueryRepository;
 import kr.modernworld.modernworldv2.user.domain.port.user.UserRepository;
@@ -11,7 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class OAuthPersistenceService {
+public class UserService {
 
   private final UserRepository userRepository;
   private final UserQueryRepository userQueryRepository;
@@ -42,5 +46,30 @@ public class OAuthPersistenceService {
     }
 
     return savedUser;
+  }
+
+  @Transactional
+  public void decreaseCurrentPoint(Long userNo, Long price) {
+    User user = userRepository.findUserByUserNoForUpdate(userNo)
+        .orElseThrow(() -> new BusinessException(BusinessErrorCode.USER_NOT_FOUND));
+
+    try {
+      user.decreaseCurrentPoint(price);
+    } catch (IllegalArgumentException e) {
+      throw new BusinessException(BusinessErrorCode.USER_NOT_HAS_ENOUGH_POINT,
+          " " + e.getMessage());
+    }
+
+    userRepository.save(user);
+  }
+
+  @Transactional
+  public void increaseCurrentAccumulationPoint(Long userNo, Long price) {
+    User user = userRepository.findUserByUserNoForUpdate(userNo)
+        .orElseThrow(() -> new BusinessException(BusinessErrorCode.USER_NOT_FOUND));
+
+    user.increaseCurrentAccumulationPoint(price);
+
+    userRepository.save(user);
   }
 }
